refactor article and megamenu component (v1)
