+++ core/src/test/java/com/aem/demo/core/components/internal/models/v1/MegaMenuImplTest.java
@@ -1,4 +1,4 @@
-package com.aem.demo.core.components.internal.models;
+package com.aem.demo.core.components.internal.models.v1;
 
 import com.aem.demo.core.components.models.MegaMenu;
 import com.aem.demo.core.components.models.MegaMenuItem;
