+++ core/src/test/resources/com/aem/demo/core/components/megamenu.json
@@ -109,7 +109,7 @@
                   "jcr:primaryType": "nt:unstructured",
                   "megamenu": {
                     "jcr:primaryType": "nt:unstructured",
-                    "sling:resourceType": "aem-demo/components/megamenu",
+                    "sling:resourceType": "aem-demo/components/megamenu/v1/megamenu",
                     "navigationRoot": "/content/aem-demo/us/en",
                     "structureDepth": 3,
                     "id": "megamenu-48ea24db16"
