+++ ui.apps/src/main/content/jcr_root/apps/aem-demo/components/article/v1/article/_cq_dialog/.content.xml
