+++ ui.apps/src/main/content/jcr_root/apps/aem-demo/components/megamenu/v1/megamenu/.content.xml
