+++ ui.config/src/main/content/jcr_root/apps/aem-demo/osgiconfig/config.publish/org.apache.sling.dynamicinclude.Configuration~aem-demo.config
@@ -1,6 +1,6 @@
 # Configuration created by Apache Sling JCR Installer
 include-filter.config.ignoreUrlParams=[""]
-include-filter.config.resource-types=["aem-demo/components/megamenu"]
+include-filter.config.resource-types=["aem-demo/components/megamenu/v1/megamenu"]
 include-filter.config.extension=""
 include-filter.config.enabled=B"true"
 include-filter.config.path="/content/experience-fragments/aem-demo"
