+++ ui.content/src/main/content/jcr_root/content/aem-demo/us/en/.content.xml
@@ -15,7 +15,7 @@
           sling:resourceType="aem-demo/components/container"
           layout="responsiveGrid">
           <artice jcr:primaryType="nt:unstructured"
-            sling:resourceType="aem-demo/components/article"
+            sling:resourceType="aem-demo/components/article/v1/article"
             articleTitle="AEM Community"
             articleDescription="&lt;p>Level 9&lt;/p>"
             textIsRich="true"
