+++ ui.content/src/main/content/jcr_root/content/experience-fragments/aem-demo/us/en/site/header/master/.content.xml
@@ -12,7 +12,7 @@
       sling:resourceType="aem-demo/components/container"
       layout="simple">
       <megamenu jcr:primaryType="nt:unstructured"
-        sling:resourceType="aem-demo/components/megamenu"
+        sling:resourceType="aem-demo/components/megamenu/v1/megamenu"
         navigationRoot="/content/aem-demo/us/en"
         structureDepth="1"/>
     </root>
